Code reorg to support webgl
