+++ Assets/_scripts/Utilities/ComponentAttributes/ComponentReferenceAttributes.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 0d98df97f9ddee44fbe38e2c4fab86b8
+guid: d252455d54fc97341bc712c2b2be5acd
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
