+++ Assets/_scripts/Utilities/ComponentAttributes/Editor.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 84b315593e39c6744b5c8906f134de65
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
