+++ Assets/_scripts/Utilities/ComponentAttributes/Editor/ComponentReferenceAttributesEditor.cs
@@ -1,32 +1,9 @@
-
+using ComponentReferenceAttributes;
 using UnityEditor;
 using UnityEngine;
 
-namespace ComponentReferenceAttributes
+namespace ComponentReferenceAttributesEditor
 {
-    public abstract class ComponentReferenceAttribute : PropertyAttribute { }
-    ///<summary>
-    /// Inspector can auto-set reference to component on same GameObject
-    ///</summary>
-    public class OwnComponentAttribute : ComponentReferenceAttribute { }
-    ///<summary>
-    /// Inspector can auto-set reference to component on sibling gameObject (siblings have same parent as this gameObject)
-    ///</summary>
-    public class SiblingComponentAttribute : ComponentReferenceAttribute { }
-    ///<summary>
-    /// Inspector can auto-set reference to component on direct parent gameobject
-    ///</summary>
-    public class ParentComponentAttribute : ComponentReferenceAttribute { }
-    ///<summary>
-    /// Inspector can auto-set reference to component on direct child gameObject
-    ///</summary>
-    public class ChildComponentAttribute : ComponentReferenceAttribute { }
-    ///<summary>
-    /// Inspector can auto-set reference to component on parent, grandparent, etc.
-    ///</summary>
-    public class AncestorComponentAttribute : ComponentReferenceAttribute { }
-    public class DescendantComponentAttribute : ComponentReferenceAttribute { }
-
     [CustomPropertyDrawer( typeof( ComponentReferenceAttribute ) )]
     [CustomPropertyDrawer( typeof( OwnComponentAttribute ) )]
     [CustomPropertyDrawer( typeof( ChildComponentAttribute ) )]
@@ -106,5 +83,4 @@ namespace ComponentReferenceAttributes
             EditorGUI.EndProperty();
         }
     }
-
 }
