+++ Assets/_scripts/Utilities/ComponentAttributes/Editor/ComponentReferenceAttributesEditor.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 18599a13e49aa7348907c0b351ab0553
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
