+++ Assets/_scripts/Utilities/Editor.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d2cee16dd23c5e04eba63d1cdad1b4fc
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
